commented out MaxBannerWidth() and changed Linker
MaxBannerWidth(in ViewReports_View.cs) was throwing a null ref
exception, so I commented that out in order to get the app to run.  I
also changed the linker setting to Link SDK assemblies only, in the iOS
build options, because it was stripping out the MvxViewModelLoader
constructor used to load the views.